cc Split valid hgroup-in-dd test into its own file
The W3C checker service treats hgroup as non-conforming everywhere. So
any conformance-checker tests containing hgroup that are expected to be
valid need to be separated out to files with "hgroup" in the filename so
that when running the conformance-checker test runner to test the W3C
service, the *hgroup*.html files can be ignored.